this works!
I'm still not sure why the limit switches aren't working as expected. this is an issue for after the ukiah competition. To allow our robot to run properly and account for jolts after releasing our second stage, spacing 4 inches away from the edge of the goal provides us with consistently satisfactory autonomous results.